# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Accounts Service API",
    "description": "Accounts API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9001",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/accounts/register": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "getAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "account-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddressModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "PhoneModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "countrycode": {
            "type": "string"
          }
        }
      },
      "RegistrationRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "pattern": "^(?=.{8,}$)(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[^A-Za-z0-9])\\S+$",
            "type": "string"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneModel"
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          }
        }
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "UpdateAccountRequest": {
        "type": "object",
        "properties": {
          "phone": {
            "$ref": "#/components/schemas/PhoneModel"
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          }
        }
      },
      "AccountModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneModel"
          },
          "address": {
            "$ref": "#/components/schemas/AddressModel"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter your JWT token in the format: **Bearer &lt;token&gt;**",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}